refactor: component name
