+++ apps/journeys-admin/src/components/Editor/Slider/Settings/CanvasDetails/Footer/Host/Host.stories.tsx
@@ -21,12 +21,12 @@ import { GET_USER_TEAMS_AND_INVITES } from '../../../../../../../libs/useUserTea
 import { ThemeProvider } from '../../../../../../ThemeProvider'
 import { DRAWER_WIDTH } from '../../../../../constants'
 
-import { GET_ALL_TEAM_HOSTS, HostTab } from './HostTab'
+import { GET_ALL_TEAM_HOSTS, Host } from './Host'
 
-const Demo: Meta<typeof HostTab> = {
+const Demo: Meta<typeof Host> = {
   ...journeysAdminConfig,
-  component: HostTab,
-  title: 'Journeys-Admin/Editor/Slider/Settings/CanvasDetails/Footer/HostTab'
+  component: Host,
+  title: 'Journeys-Admin/Editor/Slider/Settings/CanvasDetails/Footer/Host'
 }
 
 const user = {
@@ -151,14 +151,14 @@ const getTeamHostsMock: MockedResponse<
 }
 
 const Template: StoryObj<
-  ComponentProps<typeof HostTab> & { mocks: MockedResponse[] }
+  ComponentProps<typeof Host> & { mocks: MockedResponse[] }
 > = {
   render: ({ mocks, ...args }) => (
     <MockedProvider mocks={mocks}>
       <ThemeProvider>
         <JourneyProvider value={{ ...args, variant: 'admin' }}>
           <Box sx={{ width: DRAWER_WIDTH }}>
-            <HostTab />
+            <Host />
           </Box>
         </JourneyProvider>
       </ThemeProvider>
@@ -166,7 +166,7 @@ const Template: StoryObj<
   )
 }
 
-// Default HostTabs based on host availability
+// Default Hosts based on host availability
 export const Default = {
   ...Template,
   args: {
