+++ apps/journeys-admin/src/components/Editor/Slider/Settings/CanvasDetails/Footer/Host/HostForm/HostAvatarsButton/HostAvatarsButton.spec.tsx
