+++ apps/journeys-admin/src/components/Editor/Slider/Settings/CanvasDetails/Footer/Host/HostInfo/index.ts
