+++ apps/journeys-admin/src/components/Editor/Slider/Settings/CanvasDetails/Footer/Host/HostList/HostListItem/HostListItem.stories.tsx
@@ -12,7 +12,7 @@ const HostListItemDemo: Meta<typeof HostListItem> = {
   ...journeysAdminConfig,
   component: HostListItem,
   title:
-    'Journeys-Admin/Editor/Slider/Settings/CanvasDetails/Footer/HostTab/HostList/HostListItem',
+    'Journeys-Admin/Editor/Slider/Settings/CanvasDetails/Footer/Host/HostList/HostListItem',
   parameters: {
     ...journeysAdminConfig.parameters,
     layout: 'fullscreen'
