+++ apps/journeys-admin/src/components/Editor/Slider/Settings/CanvasDetails/Footer/Host/HostSelection/HostSelection.stories.tsx
@@ -9,16 +9,40 @@ import { JourneyFields as Journey } from '@core/journeys/ui/JourneyProvider/__ge
 import { simpleComponentConfig } from '@core/shared/ui/storybook'
 
 import { GetAllTeamHosts_hosts as Host } from '../../../../../../../../../__generated__/GetAllTeamHosts'
+import {
+  GetUserTeamsAndInvites,
+  GetUserTeamsAndInvites_userTeams as UserTeam
+} from '../../../../../../../../../__generated__/GetUserTeamsAndInvites'
+import { UserTeamRole } from '../../../../../../../../../__generated__/globalTypes'
 import { ThemeProvider } from '../../../../../../../ThemeProvider'
 import { DRAWER_WIDTH } from '../../../../../../constants'
 
-import { HostForm } from './HostForm'
+import { HostSelection } from './HostSelection'
 
-const Demo: Meta<typeof HostForm> = {
+const Demo: Meta<typeof HostSelection> = {
   ...simpleComponentConfig,
-  component: HostForm,
+  component: HostSelection,
   title:
-    'Journeys-Admin/Editor/Slider/Settings/CanvasDetails/Footer/HostTab/HostForm'
+    'Journeys-Admin/Editor/Slider/Settings/CanvasDetails/Footer/Host/HostSelection'
+}
+
+const user = {
+  id: 'userId',
+  email: 'admin@email.com'
+}
+
+const userTeam: UserTeam = {
+  id: 'teamId',
+  __typename: 'UserTeam',
+  role: UserTeamRole.manager,
+  user: {
+    __typename: 'User',
+    email: user.email,
+    firstName: 'User',
+    id: user.id,
+    imageUrl: 'imageURL',
+    lastName: '1'
+  }
 }
 
 const defaultHost: Host = {
@@ -35,7 +59,7 @@ const journey = {
   id: 'journeyId',
   seoTitle: 'My awesome journey',
   host: defaultHost,
-  team: { id: 'teamId', title: 'My team' },
+  team: { id: userTeam.id, title: 'My team' },
   language: {
     __typename: 'Language',
     id: '529',
@@ -49,9 +73,16 @@ const journey = {
   }
 } as unknown as Journey
 
+const data: GetUserTeamsAndInvites = {
+  userTeams: [userTeam],
+  userTeamInvites: []
+}
+
+const handleSelection = jest.fn()
+
 const Template: StoryObj<{
   journey: Journey
-  componentProps: ComponentProps<typeof HostForm>
+  componentProps: ComponentProps<typeof HostSelection>
 }> = {
   render: ({ journey, componentProps }) => {
     return (
@@ -59,7 +90,7 @@ const Template: StoryObj<{
         <ThemeProvider>
           <JourneyProvider value={{ journey, variant: 'admin' }}>
             <Box sx={{ width: DRAWER_WIDTH }}>
-              <HostForm {...componentProps} />
+              <HostSelection {...componentProps} />
             </Box>
           </JourneyProvider>
         </ThemeProvider>
@@ -73,19 +104,33 @@ export const Default = {
   args: {
     journey: { ...journey, host: null },
     componentProps: {
-      handleSelection: jest.fn(),
-      getAllTeamHosts: jest.fn()
+      userInTeam: true,
+      data,
+      handleSelection
     }
   }
 }
 
-export const Filled = {
+export const EditHost = {
   ...Template,
   args: {
     journey,
     componentProps: {
-      handleSelection: jest.fn(),
-      getAllTeamHosts: jest.fn()
+      userInTeam: true,
+      data,
+      handleSelection
+    }
+  }
+}
+
+export const Disabled = {
+  ...Template,
+  args: {
+    journey: { ...journey, host: null },
+    componentProps: {
+      userInTeam: false,
+      data,
+      handleSelection
     }
   }
 }
