+++ apps/journeys-admin/src/components/Editor/Slider/Settings/CanvasDetails/Footer/Host/index.ts
@@ -1,0 +1,1 @@
+export { Host } from './Host'
