--- apps/journeys-admin/src/components/Editor/Slider/Settings/CanvasDetails/Footer/HostTab/index.ts
@@ -1,1 +1,0 @@
-export { HostTab } from './HostTab'
